+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.0.3
+
+- Require a comma at the end ob multiline arrays, objects, imports, ...
+- Ignore js files in the root directory because they are not part of the typescript rootDir
+
 ## v0.0.2
 
 - Disabled rule `@typescript-eslint/explicit-function-return-type`